+++ src/datasource/tc_monitor/cmq_topic/query_def.ts
@@ -67,11 +67,36 @@ export const InvalidDemensions = {
 // 因为cmq那边，以TopciName为主键
 export const templateQueryIdMap = {
   instance: 'TopicName',
+  subscription: 'SubscriptionId',
 };
 
+// 需要缓存到storage的内容的key列表
+const keyInStorage = {
+  subscriptionId: 'subscriptionId',
+};
+/*
+如果有InstanceId额外的维度，原则上都需要传入此map结构配置
+key的含义：
+  经过InvalidDemensions处理后的string。topicId =》TopicId。
+  否则认为指标中维度正确，和指标中维度字段保持一致，即topicId
+value的含义：
+  1 dim_KeyInStorage 指标中维度dimension对应的storage中的key，获取缓存列表，sourceMapList、
+  2 dim_KeyInTarget  通过getVariable方法获取变量中选中项。比如ListnerId为Lis-xxxx；即：STATE中的key。
+                    默认取通过InvalidDemsion处理后的key
+  3 dim_KeyInMap     保存在模板变量value比如（监听器ID）源自sourceMapList（接口返回内容）的哪个key（ListenerId）。
+                    即：templateQueryIdMap中内容。
+                    联合上面2的内容筛选出原始sourceMap
+*/
+const queryMonitorExtraConfg = {
+  subscriptionId: {
+    dim_KeyInStorage: keyInStorage.subscriptionId,
+    dim_KeyInTarget: 'subscriptionId',
+    dim_KeyInMap: templateQueryIdMap.subscription,
+  },
+};
 // 与名称挂钩的变量
 const InstanceAliasList = ['TopicName', 'TopicId'];
-export { InstanceAliasList, GetInstanceQueryParams };
+export { InstanceAliasList, GetInstanceQueryParams, queryMonitorExtraConfg, keyInStorage };
 
 // 界面状态模型，default
 export default {