+++ src/datasource/tc_monitor/ecm_block_storage/query_def.ts
@@ -9,6 +9,7 @@ const queryEditorName = 'ecmBlockStorageQuery';
 
 const ECMBLOCKSTORAGEInvalidDemensions = {
   uuid: 'UUID',
+  diskId: 'DiskId',
 };
 
 // 需和文件名对应
@@ -16,6 +17,7 @@ const ECMBLOCKSTORAGEInstanceAliasList = ['InstanceId', 'InstanceName'];
 
 const templateQueryIdMap = {
   instance: 'InstanceId',
+  diskId: 'DiskId',
 };
 
 // select类型需要注意是{},multi后缀是[],dropdown是''
@@ -179,6 +181,31 @@ const ECMBLOCKSTORAGE_STATE = {
   queries: ECMBLOCKSTORAGEFilterFields,
 };
 
+// 需要缓存到storage的内容的key列表
+const keyInStorage = {
+  diskId: 'diskId',
+};
+/*
+如果有InstanceId额外的维度，原则上都需要传入此map结构配置
+key的含义：
+  经过InvalidDemensions处理后的string。topicId =》TopicId。
+  否则认为指标中维度正确，和指标中维度字段保持一致，即topicId
+value的含义：
+  1 dim_KeyInStorage 指标中维度dimension对应的storage中的key，获取缓存列表，sourceMapList、
+  2 dim_KeyInTarget  通过getVariable方法获取变量中选中项。比如ListnerId为Lis-xxxx；即：STATE中的key。
+                    默认取通过InvalidDemsion处理后的key
+  3 dim_KeyInMap     保存在模板变量value比如（监听器ID）源自sourceMapList（接口返回内容）的哪个key（ListenerId）。
+                    即：templateQueryIdMap中内容。
+                    联合上面2的内容筛选出原始sourceMap
+*/
+const queryMonitorExtraConfg = {
+  DiskId: {
+    dim_KeyInStorage: keyInStorage.diskId,
+    dim_KeyInTarget: 'diskId',
+    dim_KeyInMap: templateQueryIdMap.diskId,
+  },
+};
+
 function GetInstanceQueryParams(queries: any = {}) {
   return instanceQueryParamsBaseParse(queries, true);
 }
@@ -191,6 +218,8 @@ export {
   namespace,
   queryEditorName,
   queryEditorConfig,
+  keyInStorage,
+  queryMonitorExtraConfg,
   // 对应产品的service的全大写拼接GetInstanceQueryParams
   GetInstanceQueryParams as ECMBLOCKSTORAGEGetInstanceQueryParams,
 };