+++ CHANGELOG.md
@@ -1,6 +1,28 @@
 # TencentCloud Monitor Grafana App
 ---
 
+## [2.2.2] - 2021-08-27
+
+### Fixed
+- Legend name display optimization, support multi-level display.
+- Support `topicId` template variable in CKafka monitoring.
+- Support select instance through `As WanIp` in Redis monitoring.
+- Support template variable `fleet` and `queue` in Game Server Elastic-scaling monitoring, `InstanceAlias` support `PrivateIpAddress` and `IpAddress`.
+
+See detailed guide in [README.md](https://github.com/TencentCloud/tencentcloud-monitor-grafana-app/blob/master/README.md).
+
+---
+
+## [2.2.1] - 2021-07-13
+
+### Fixed
+- Missing dimensions in lbPrivate.
+- Missing metrics `BaseCpuUsage`, `CvmDiskUsage` for CVM monitoring.
+
+See detailed guide in [README.md](https://github.com/TencentCloud/tencentcloud-monitor-grafana-app/blob/master/README.md).
+
+---
+
 ## [2.2.0] - 2021-06-28
 
 ### Added